FL-102 delete 시 repository에서 최근 id를 받도록 수정 (화면 연동 부분은 없음)

+++ flamingo-web/src/main/java/org/exem/flamingo/web/oozie/workflow/OozieWorkflowRepository.java
@@ -14,5 +14,5 @@ public interface OozieWorkflowRepository {
   public Map getRecentWorkflow();
   public void insertWorkflow(Map param);
   public void updateWorkflow(Map param);
-  public void deleteWorkflow(int id);
+  public void deleteWorkflow(long id);
 }
